backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
supabase
tiktoken
pymupdf
python-multipart
cohere
numpy
